settings splited to fragment
