+++ backend/hct_mis_api/apps/core/celery.py
@@ -7,7 +7,7 @@ CELERY_QUEUE_DEFAULT = "default"
 CELERY_QUEUE_OTHER = "priority"
 
 # set the default Django settings module for the 'celery' program.
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "hct_mis_api.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "hct_mis_api.settings.settings")
 
 app = Celery("hct_mis")
