+++ backend/hct_mis_api/apps/core/field_attributes/core_fields_attributes.py
@@ -184,6 +184,7 @@ CORE_FIELDS_ATTRIBUTES = [
         "choices": [],
         "associated_with": _HOUSEHOLD,
         "xlsx_field": "address_h_c",
+        "core_field_required_by_deliver_mechanism": ['TRANSFER'],
         "scope": [Scope.GLOBAL, Scope.TARGETING, Scope.KOBO_IMPORT, Scope.HOUSEHOLD_UPDATE],
     },
     {
