+++ backend/hct_mis_api/settings/fragments/loggers.py
@@ -1,0 +1,51 @@
+import logging
+import sys
+from typing import Dict, Any
+
+from hct_mis_api.settings.env import env
+from hct_mis_api.settings.settings import DEBUG
+
+LOG_LEVEL = env("LOG_LEVEL", default="ERROR") if "test" not in sys.argv else "INFO"
+
+LOGGING: Dict[str, Any] = {
+    "version": 1,
+    "disable_existing_loggers": False,
+    "formatters": {
+        "standard": {"format": "%(asctime)s [%(levelname)s] %(name)s line %(lineno)d: %(message)s"},
+        "verbose": {
+            "format": "[%(asctime)s][%(levelname)s][%(name)s] %(filename)s.%(funcName)s:%(lineno)d %(message)s",
+        },
+    },
+    "filters": {"require_debug_false": {"()": "django.utils.log.RequireDebugFalse"}},
+    "handlers": {
+        "default": {
+            "level": LOG_LEVEL,
+            "class": "logging.StreamHandler",
+            "formatter": "standard",
+        },
+        "file": {
+            "level": LOG_LEVEL,
+            "class": "logging.FileHandler",
+            "filename": "debug.log",
+        },
+    },
+    "loggers": {
+        "": {"handlers": ["default"], "level": "INFO", "propagate": True},
+        "console": {"handlers": ["default"], "level": "DEBUG", "propagate": True},
+        "django.request": {
+            "handlers": ["default"],
+            "level": "DEBUG" if DEBUG else "ERROR",
+            "propagate": False,
+        },
+        "django.security.DisallowedHost": {
+            # Skip "SuspiciousOperation: Invalid HTTP_HOST" e-mails.
+            "handlers": ["default"],
+            "propagate": False,
+        },
+        "elasticsearch": {"handlers": ["file"], "level": "CRITICAL", "propagate": True},
+    },
+}
+
+# overwrite Azure logs
+logger_azure = logging.getLogger("azure.core.pipeline.policies.http_logging_policy")
+logger_azure.setLevel(logging.WARNING)
