+++ backend/hct_mis_api/settings/fragments/mailjet.py
@@ -1,0 +1,4 @@
+from hct_mis_api.settings.env import env
+
+MAILJET_API_KEY = env("MAILJET_API_KEY")
+MAILJET_SECRET_KEY = env("MAILJET_SECRET_KEY")
