+++ backend/hct_mis_api/settings/settings_test.py
@@ -1,7 +1,7 @@
 import logging
 import os
 
-from .settings import *  # noqa: ignore=F403
+from backend.hct_mis_api.settings import *  # noqa: ignore=F403
 
 # dev overrides
 DEBUG = True
